Added the delete function for the Calculator
