--- src/main/java/ie/atu/calculator/CalculatorDelete.java
@@ -1,21 +1,0 @@
-package ie.atu.calculator;
-import java.util.Scanner;
-
-    public class CalculatorDelete {
-
-        public static void delete(){
-
-            Scanner inputs = new Scanner(System.in);
-
-            System.out.println("Enter your first numbers");
-            int firstNumber = inputs.nextInt();
-
-            System.out.println("Enter your second numbers");
-            int secondNumber = inputs.nextInt();
-
-            int total = firstNumber - secondNumber;
-            System.out.println("The total is: "+total);
-
-        }
-    }
-
